switch usage of float to double basically everywhere; fix -Wnarrowing warnings by casting

+++ src/character.hpp
@@ -9,11 +9,11 @@
 class Character : Entity{
 public:
     Character(const GridCellIndex &pos, const Angle &facing_angle);
-    void Draw(const DrawPosition &grid_tl_pos, float grid_cell_width, float grid_cell_height) override;
+    void Draw(const DrawPosition &grid_tl_pos, double grid_cell_width, double grid_cell_height) override;
     void Update(int fps) override;
 private:
     Angle facing_angle;
-    float command_cooldown;
+    double command_cooldown;
 };
 
 #endif // CHARACTER_HPP
