+++ src/entity.hpp
@@ -15,7 +15,7 @@ enum class EntityAlignment {
 class Entity {
 public:
     Entity(EntityAlignment alignment, std::variant<GridCellIndex, GridContinuousPosition> pos, const Movement &movement = Movement(AngleType::Continuous, std::variant<Wind8MovementInfo, ContinuousAngle>(0.0f), 0.0), WallCollisionBehavior wcb = WallCollisionBehavior::Ignore) : alignment(alignment), pos(pos), movement(movement), wcb(wcb) {}
-    virtual void Draw(const DrawPosition &grid_tl_pos, float grid_cell_width, float grid_cell_height) = 0;
+    virtual void Draw(const DrawPosition &grid_tl_pos, double grid_cell_width, double grid_cell_height) = 0;
     void Move(const GridCellIndex &grid_dimensions, int fps);
     virtual void Update(int fps) = 0; // called each frame
     EntityAlignment alignment;
